automod whitelisting + commands split
