+++ src/commands/admin/automod/anti.js
@@ -1,0 +1,178 @@
+/**
+ * @type {import("@structures/Command")}
+ */
+module.exports = {
+  name: "anti",
+  description: "manage various automod settings for the server",
+  category: "AUTOMOD",
+  userPermissions: ["MANAGE_GUILD"],
+  command: {
+    enabled: true,
+    minArgsCount: 2,
+    subcommands: [
+      {
+        trigger: "ghostping <on|off>",
+        description: "detect and logs ghost mentions in your server",
+      },
+      {
+        trigger: "spam <on|off>",
+        description: "enable or disable antispam detection",
+      },
+      {
+        trigger: "massmention <on|off> [threshold]",
+        description: "enable or disable massmention detection [default threshold is 3 mentions]",
+      },
+    ],
+  },
+  slashCommand: {
+    enabled: true,
+    ephemeral: true,
+    options: [
+      {
+        name: "ghostping",
+        description: "detects and logs ghost mentions in your server",
+        type: "SUB_COMMAND",
+        options: [
+          {
+            name: "status",
+            description: "configuration status",
+            required: true,
+            type: "STRING",
+            choices: [
+              {
+                name: "ON",
+                value: "ON",
+              },
+              {
+                name: "OFF",
+                value: "OFF",
+              },
+            ],
+          },
+        ],
+      },
+      {
+        name: "spam",
+        description: "enable or disable antispam detection",
+        type: "SUB_COMMAND",
+        options: [
+          {
+            name: "status",
+            description: "configuration status",
+            required: true,
+            type: "STRING",
+            choices: [
+              {
+                name: "ON",
+                value: "ON",
+              },
+              {
+                name: "OFF",
+                value: "OFF",
+              },
+            ],
+          },
+        ],
+      },
+      {
+        name: "massmention",
+        description: "enable or disable massmention detection",
+        type: "SUB_COMMAND",
+        options: [
+          {
+            name: "status",
+            description: "configuration status",
+            required: true,
+            type: "STRING",
+            choices: [
+              {
+                name: "ON",
+                value: "ON",
+              },
+              {
+                name: "OFF",
+                value: "OFF",
+              },
+            ],
+          },
+          {
+            name: "threshold",
+            description: "configuration threshold (default is 3 mentions)",
+            required: false,
+            type: "INTEGER",
+          },
+        ],
+      },
+    ],
+  },
+
+  async messageRun(message, args, data) {
+    const settings = data.settings;
+    const sub = args[0].toLowerCase();
+
+    let response;
+    if (sub == "ghostping") {
+      const status = args[1].toLowerCase();
+      if (!["on", "off"].includes(status)) return message.safeReply("Invalid status. Value must be `on/off`");
+      response = await antiGhostPing(settings, status);
+    }
+
+    //
+    else if (sub == "spam") {
+      const status = args[1].toLowerCase();
+      if (!["on", "off"].includes(status)) return message.safeReply("Invalid status. Value must be `on/off`");
+      response = await antiSpam(settings, status);
+    }
+
+    //
+    else if (sub === "massmention") {
+      const status = args[1].toLowerCase();
+      const threshold = args[2] || 3;
+      if (!["on", "off"].includes(status)) return message.safeReply("Invalid status. Value must be `on/off`");
+      response = await antiMassMention(settings, status, threshold);
+    }
+
+    //
+    else response = "Invalid command usage!";
+    await message.safeReply(response);
+  },
+
+  async interactionRun(interaction, data) {
+    const sub = interaction.options.getSubcommand();
+    const settings = data.settings;
+
+    let response;
+    if (sub == "ghostping") response = await antiGhostPing(settings, interaction.options.getString("status"));
+    else if (sub == "spam") response = await antiSpam(settings, interaction.options.getString("status"));
+    else if (sub === "massmention") {
+      response = await antiMassMention(settings, interaction.options.getInteger("amount"));
+    } else response = "Invalid command usage!";
+
+    await interaction.followUp(response);
+  },
+};
+
+async function antiGhostPing(settings, input) {
+  const status = input.toUpperCase() === "ON" ? true : false;
+  settings.automod.anti_ghostping = status;
+  await settings.save();
+  return `Configuration saved! Anti-Ghostping is now ${status ? "enabled" : "disabled"}`;
+}
+
+async function antiSpam(settings, input) {
+  const status = input.toUpperCase() === "ON" ? true : false;
+  settings.automod.anti_spam = status;
+  await settings.save();
+  return `Antispam detection is now ${status ? "enabled" : "disabled"}`;
+}
+
+async function antiMassMention(settings, input, threshold) {
+  const status = input.toUpperCase() === "ON" ? true : false;
+  if (!status) {
+    settings.automod.anti_massmention = 0;
+  } else {
+    settings.automod.anti_massmention = threshold;
+  }
+  await settings.save();
+  return `Mass mention detection is now ${status ? "enabled" : "disabled"}`;
+}
