--- src/commands/admin/automod/automodconfig.js
@@ -1,221 +1,0 @@
-const { MessageEmbed } = require("discord.js");
-const { EMBED_COLORS } = require("@root/config.js");
-const { stripIndent } = require("common-tags");
-
-/**
- * @type {import("@structures/Command")}
- */
-module.exports = {
-  name: "automodconfig",
-  description: "various automod configuration",
-  category: "AUTOMOD",
-  userPermissions: ["MANAGE_GUILD"],
-  command: {
-    enabled: true,
-    minArgsCount: 1,
-    subcommands: [
-      {
-        trigger: "status",
-        description: "check automod configuration for this guild",
-      },
-      {
-        trigger: "strikes <number>",
-        description: "maximum number of strikes a member can receive before taking an action",
-      },
-      {
-        trigger: "action <TIMEOUT|KICK|BAN>",
-        description: "set action to be performed after receiving maximum strikes",
-      },
-      {
-        trigger: "debug <on|off>",
-        description: "turns on automod for messages sent by admins and moderators",
-      },
-    ],
-  },
-  slashCommand: {
-    enabled: true,
-    ephemeral: true,
-    options: [
-      {
-        name: "status",
-        description: "Check automod configuration",
-        type: "SUB_COMMAND",
-      },
-      {
-        name: "strikes",
-        description: "Set maximum number of strikes before taking an action",
-        type: "SUB_COMMAND",
-        options: [
-          {
-            name: "amount",
-            description: "number of strikes (default 5)",
-            required: true,
-            type: "INTEGER",
-          },
-        ],
-      },
-      {
-        name: "action",
-        description: "Set action to be performed after receiving maximum strikes",
-        type: "SUB_COMMAND",
-        options: [
-          {
-            name: "action",
-            description: "action to perform",
-            type: "STRING",
-            required: true,
-            choices: [
-              {
-                name: "TIMEOUT",
-                value: "TIMEOUT",
-              },
-              {
-                name: "KICK",
-                value: "KICK",
-              },
-              {
-                name: "BAN",
-                value: "BAN",
-              },
-            ],
-          },
-        ],
-      },
-      {
-        name: "debug",
-        description: "Enable/disable automod for messages sent by admins & moderators",
-        type: "SUB_COMMAND",
-        options: [
-          {
-            name: "status",
-            description: "configuration status",
-            required: true,
-            type: "STRING",
-            choices: [
-              {
-                name: "ON",
-                value: "ON",
-              },
-              {
-                name: "OFF",
-                value: "OFF",
-              },
-            ],
-          },
-        ],
-      },
-    ],
-  },
-
-  async messageRun(message, args, data) {
-    const input = args[0].toLowerCase();
-    const settings = data.settings;
-
-    let response;
-    if (input === "status") {
-      response = await getStatus(settings, message.guild);
-    } else if (input === "strikes") {
-      const strikes = args[1];
-      if (isNaN(strikes) || Number.parseInt(strikes) < 1) {
-        return message.safeReply("Strikes must be a valid number greater than 0");
-      }
-      response = await setStrikes(settings, strikes);
-    } else if (input === "action") {
-      const action = args[1].toUpperCase();
-      if (!action || !["TIMEOUT", "KICK", "BAN"].includes(action))
-        return message.safeReply("Not a valid action. Action can be `Timeout`/`Kick`/`Ban`");
-      response = await setAction(settings, message.guild, action);
-    } else if (input === "debug") {
-      const status = args[1].toLowerCase();
-      if (!["on", "off"].includes(status)) return message.safeReply("Invalid status. Value must be `on/off`");
-      response = await setDebug(settings, status);
-    }
-
-    //
-    else response = "Invalid command usage!";
-    await message.safeReply(response);
-  },
-
-  async interactionRun(interaction, data) {
-    const sub = interaction.options.getSubcommand();
-    const settings = data.settings;
-
-    let response;
-
-    // status
-    if (sub === "status") response = await getStatus(settings, interaction.guild);
-    else if (sub === "strikes") response = await setStrikes(settings, interaction.options.getInteger("amount"));
-    else if (sub === "action")
-      response = await setAction(settings, interaction.guild, interaction.options.getString("action"));
-    else if (sub === "debug") response = await setDebug(settings, interaction.options.getString("status"));
-
-    await interaction.followUp(response);
-  },
-};
-
-async function getStatus(settings, guild) {
-  const { automod } = settings;
-
-  const logChannel = settings.modlog_channel
-    ? guild.channels.cache.get(settings.modlog_channel).toString()
-    : "Not Configured";
-
-  // String Builder
-  let desc = stripIndent`
-    ❯ **Max Lines**: ${automod.max_lines || "NA"}
-    ❯ **Anti-Massmention**: ${automod.anti_massmention > 0 ? "✓" : "✕"}
-    ❯ **Anti-Attachment**: ${automod.anti_attachment ? "✓" : "✕"}
-    ❯ **Anti-Links**: ${automod.anti_links ? "✓" : "✕"}
-    ❯ **Anti-Invites**: ${automod.anti_invites ? "✓" : "✕"}
-    ❯ **Anti-Spam**: ${automod.anti_spam ? "✓" : "✕"}
-    ❯ **Anti-Ghostping**: ${automod.anti_ghostping ? "✓" : "✕"}
-  `;
-
-  const embed = new MessageEmbed()
-    .setAuthor({ name: "Automod Configuration", icon_url: guild.iconURL() })
-    .setColor(EMBED_COLORS.BOT_EMBED)
-    .setDescription(desc)
-    .addField("Log Channel", logChannel, true)
-    .addField("Max Strikes", automod.strikes.toString(), true)
-    .addField("Action", automod.action, true)
-    .addField("Debug", automod.debug ? "✓" : "✕", true);
-
-  return { embeds: [embed] };
-}
-
-async function setStrikes(settings, strikes) {
-  settings.automod.strikes = strikes;
-  await settings.save();
-  return `Configuration saved! Maximum strikes is set to ${strikes}`;
-}
-
-async function setAction(settings, guild, action) {
-  if (action === "TIMEOUT") {
-    if (!guild.me.permissions.has("MODERATE_MEMBERS")) {
-      return "I do not permission to timeout members";
-    }
-  }
-
-  if (action === "KICK") {
-    if (!guild.me.permissions.has("KICK_MEMBERS")) {
-      return "I do not have permission to kick members";
-    }
-  }
-
-  if (action === "BAN") {
-    if (!guild.me.permissions.has("BAN_MEMBERS")) {
-      return "I do not have permission to ban members";
-    }
-  }
-
-  settings.automod.action = action;
-  await settings.save();
-  return `Configuration saved! Automod action is set to ${action}`;
-}
-
-async function setDebug(settings, input) {
-  const status = input.toLowerCase() === "on" ? true : false;
-  settings.automod.debug = status;
-  await settings.save();
-  return `Configuration saved! Automod debug is now ${status ? "enabled" : "disabled"}`;
-}
